% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_acs.R
\name{add_acs}
\alias{add_acs}
\title{Add ACS variables
Add ACS variables to data frames.}
\usage{
add_acs(
  data,
  variable,
  geography,
  year = 2019,
  by_name = FALSE,
  CENSUS_API_KEY = Sys.getenv("CENSUS_API_KEY")
)
}
\arguments{
\item{data}{Data to add ACS variables to}

\item{variable}{ACS variable(s) to add}

\item{geography}{Geographic unit}

\item{year}{Vintage}

\item{by_name}{TRUE to use county name for matching}

\item{CENSUS_API_KEY}{Census API key}
}
\description{
Add ACS variables
Add ACS variables to data frames.
}
