% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_census.R
\name{add_census}
\alias{add_census}
\title{Add Census variables}
\usage{
add_census(
  data,
  variable,
  geography,
  year = 2010,
  by_name = FALSE,
  CENSUS_API_KEY = Sys.getenv("CENSUS_API_KEY")
)
}
\arguments{
\item{data}{Data to add Census variables to}

\item{variable}{Census variable(s) to add}

\item{geography}{Geographic unit}

\item{year}{Vintage}

\item{by_name}{TRUE to use county name for matching}

\item{CENSUS_API_KEY}{Census API key}
}
\description{
Add Census variables to data frames.
}
