% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry.R
\name{add_geometry}
\alias{add_geometry}
\title{Add geometries to Census data frames}
\usage{
add_geometry(data, geography, year = NULL)
}
\arguments{
\item{data}{Census data to attach to shapefiles}

\item{geography}{Geographic unit  of the geometry}

\item{year}{Vintage of the geographic unit. Valid inputs depend on what the geography is.}
}
\value{
An sf object
}
\description{
Add geometries to Census data frames that are imported using other functions in this package.
}
