% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_access_map.R
\name{create_access_map}
\alias{create_access_map}
\title{Create family access map}
\usage{
create_access_map(
  geography = "census-tract",
  as_of = "2021-04-01",
  access_to = "licensed_providers",
  radius = 10,
  acs_year = 2019,
  kid_per_dot = 100,
  decay = 4
)
}
\arguments{
\item{geography}{Census geography to get the centroid of}

\item{as_of}{Access as of date}

\item{access_to}{Either licensed providers or all providers in Nware. Default is licensed providers}

\item{radius}{Consider providers/families that are in \code{radius} miles or closer.}

\item{acs_year}{ACS vintage for demographic variables}

\item{kid_per_dot}{Number of kids per dot on the map}

\item{decay}{Decay parameter (used to weight access by distance) in calculating the access measure}
}
\value{
An tmap object (html)
}
\description{
Create adjusted-supply family access map
}
