% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexagons.R
\name{create_hexagon_access_map}
\alias{create_hexagon_access_map}
\title{Create access map using hexagons}
\usage{
create_hexagon_access_map(
  geography = "census-block-group",
  geography_year = 2010,
  acs_year = 2019,
  hex_side = 1,
  radius = 10,
  decay = 4,
  kid_per_dot = 100,
  as_of = "2021-04-01"
)
}
\arguments{
\item{geography}{Geography}

\item{geography_year}{Geography year. 2010 or 2020}

\item{acs_year}{5-year ACS end year}

\item{hex_side}{Length of side of hexagons}

\item{radius}{Consider providers/families that are in \code{radius} miles or closer.}

\item{decay}{Decay parameter (used to weight access by distance) in calculating the access measure}

\item{kid_per_dot}{Number of kids per dot on the map}

\item{as_of}{Access as of date}
}
\value{
A tmap object
}
\description{
Create child care access map using hexagons
}
