% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs.R
\name{get_acs}
\alias{get_acs}
\title{Get ACS Data}
\usage{
get_acs(
  variable,
  geography,
  year = 2019,
  CENSUS_API_KEY = Sys.getenv("CENSUS_API_KEY")
)
}
\arguments{
\item{variable}{ACS variable}

\item{geography}{Geographic unit for the data}

\item{year}{ACS 5-year vintage (end year)}

\item{CENSUS_API_KEY}{Census API Key}
}
\value{
A tibble
}
\description{
Get ACS data about population, income, etc.
}
