% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{Get Decennial Census Data}
\usage{
get_census(
  variable,
  geography,
  year = 2010,
  CENSUS_API_KEY = Sys.getenv("CENSUS_API_KEY")
)
}
\arguments{
\item{variable}{Census variable}

\item{geography}{Geographic unit for the data}

\item{year}{Census vintage}

\item{CENSUS_API_KEY}{Census API Key}
}
\value{
A tibble of Census data.
}
\description{
Get Decennial Census data about population, etc.
}
