% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroids.R
\name{get_centroids}
\alias{get_centroids}
\title{Get population-weighted centroids for Census geographies}
\usage{
get_centroids(geography, year = 2010)
}
\arguments{
\item{geography}{Geographical unit}

\item{year}{Vintage}
}
\description{
Get centroids for Census geographies like census block group, census tract, and for some other geographies.
}
