% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_to_providers.R
\name{get_distance_to_providers}
\alias{get_distance_to_providers}
\title{Get Euclidean  distance}
\usage{
get_distance_to_providers(
  geography = "census-tract",
  year = 2010,
  destination = "licensing",
  as_of = "2021-04-01"
)
}
\arguments{
\item{geography}{Geographies to add to. Defaults to all geographies available.}

\item{year}{Vintage for the geography}

\item{destination}{List of providers to calculate distance to}

\item{as_of}{Date}
}
\value{
A tibble
}
\description{
Get Euclidean distance from population-weighted centroids to child care providers
}
