% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drive_time.R
\name{get_drive_time}
\alias{get_drive_time}
\title{Get drive time}
\usage{
get_drive_time(data, max_distance = 30)
}
\arguments{
\item{data}{A data frame with a pair of locations. Columns \code{id}, \code{lat}, \code{lon}, \code{id2}, \code{lat2}, \code{lon2} should be present.}

\item{max_distance}{Maximum distance to calculate drive times}
}
\value{
A tibble
}
\description{
Get drive time for pairs of locations.
}
