% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{get_geometry}
\alias{get_geometry}
\title{Read shapefiles.}
\usage{
get_geometry(geography, year = NULL, GDRIVE_ROOT = Sys.getenv("GDRIVE_ROOT"))
}
\arguments{
\item{geography}{Geographic unit to map. Can be "census-block", "census-block-group", "census-tract", "county", etc.}

\item{year}{Vintage of the geographic unit. Valid inputs depend on what the geography is.}

\item{GDRIVE_ROOT}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
An sf object
}
\description{
Read shapefiles from the shared Google Drive folder.
}
