% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_geometry.R
\name{import_geometry}
\alias{import_geometry}
\title{Read shapefiles.}
\usage{
import_geometry(geography, year = NULL, gdrive_root = "~/Google Drive")
}
\arguments{
\item{geography}{Geographic unit to map. Can be "census-block", "census-block-group", "census-tract", "county", etc.}

\item{year}{Vintage of the geographic unit. Valid inputs depend on what the geography is.}

\item{gdrive_root}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
An sf object
}
\description{
Read shapefiles from the shared Google Drive folder.
}
