% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_access.R
\name{measure_access}
\alias{measure_access}
\title{Measure access}
\usage{
measure_access(
  geography = "census-tract",
  year = 2010,
  as_of = "2021-04-01",
  acs_year = 2019,
  radius = 10,
  decay = 4
)
}
\arguments{
\item{geography}{Geography}

\item{year}{Year}

\item{as_of}{Date}

\item{acs_year}{ACS year}

\item{radius}{Mile radius around geography centroid}

\item{decay}{Decay parameter for weighting function}
}
\description{
Measure access of families to child care
}
