% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_access_census.R
\name{measure_access_census}
\alias{measure_access_census}
\title{Measure child care access using census geometries}
\usage{
measure_access_census(
  geography = "census-block-group",
  geo_year = 2010,
  as_of = "2021-04-01",
  acs_year = 2019,
  radius = 10,
  decay = 4
)
}
\arguments{
\item{geography}{Geography}

\item{geo_year}{Geography vintage}

\item{as_of}{Date}

\item{acs_year}{ACS year}

\item{radius}{Mile radius around geography centroid}

\item{decay}{Decay parameter for weighting function}
}
\description{
Measure access of families to child care using census geographies
}
