% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_access_hexagon.R
\name{measure_access_hexagon}
\alias{measure_access_hexagon}
\title{Measure child care access using hexagons}
\usage{
measure_access_hexagon(
  geography = "census-block",
  geo_year = 2010,
  acs_year = 2019,
  hex_side = 1,
  radius = 10,
  decay = 4,
  as_of = "2021-04-01"
)
}
\arguments{
\item{geography}{Geography}

\item{geo_year}{Geography year. 2010 or 2020}

\item{acs_year}{5-year ACS end year}

\item{hex_side}{Length of side of hexagons}

\item{radius}{Consider providers/families that are in \code{radius} miles or closer.}

\item{decay}{Decay parameter (used to weight access by distance) in calculating the access measure}

\item{as_of}{Access as of date}
}
\value{
An sf object
}
\description{
Measure access of families to child care using hexagons
}
