% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot shapefiles}
\usage{
plot_map(
  data,
  method = "tmap",
  fill_with,
  tmap_mode = "view",
  alpha = 0.5,
  ...
)
}
\arguments{
\item{data}{An sf object that contains the geometries}

\item{method}{Plotting method. Can be "sf" or "tmap"}

\item{fill_with}{(Optional) The variable from the sf object to fill the polygons.}

\item{tmap_mode}{Map mode if tmap is chosen as the mapping method.}

\item{alpha}{Transparency}

\item{...}{Other arguments passed on to \code{geom_sf}, \code{tmap_polygons}, and \code{tmap_dots}}
}
\value{
A ggplot or tmap object.
}
\description{
Plot geography shapefiles using geom_sf
}
