% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licensing.R
\name{read_licensing}
\alias{read_licensing}
\title{Read DHS Licensing Data}
\usage{
read_licensing(table = "provider", version = "2021-04", filetype = "csv")
}
\arguments{
\item{table}{character}

\item{version}{character}

\item{filetype}{character}
}
\value{
tibble
}
\description{
Read DHS Licensing Data that lives in the shared Google Drive folder.
}
