% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nware.R
\name{read_nware}
\alias{read_nware}
\title{Read Nware Data}
\usage{
read_nware(table = "characteristics", version = "2021-04", filetype = "csv")
}
\arguments{
\item{table}{character}

\item{version}{character}

\item{filetype}{character}
}
\value{
tibble
}
\description{
Read Nware Data that lives in the shared Google Drive folder.
}
