% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pecc.R
\name{read_pecc}
\alias{read_pecc}
\title{Read PECC Grant Data}
\usage{
read_pecc(filetype = "rds", GDRIVE_ROOT = Sys.getenv("GDRIVE_ROOT"))
}
\arguments{
\item{filetype}{"rds" or "csv"}

\item{GDRIVE_ROOT}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
A tibble
}
\description{
Read PECC Grant Data from the shared Google Drive folder.
}
