% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_public.R
\name{read_public}
\alias{read_public}
\title{Read Public School Enrollment Data}
\source{
Minnesota Department of Education
}
\usage{
read_public(
  table = "school",
  total = TRUE,
  filetype = "rds",
  gdrive_root = "~/Google Drive"
)
}
\arguments{
\item{table}{"school" or "district"}

\item{total}{\code{TRUE} to ignore demographic subgroups.}

\item{filetype}{"rds" or "csv"}

\item{gdrive_root}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
A tibble
}
\description{
Read Public School Enrollment Data from the shared Google Drive folder.
}
