% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_families.R
\name{simulate_families}
\alias{simulate_families}
\title{Create synthetic locations for families}
\usage{
simulate_families(data, kid_per_dot = 4)
}
\arguments{
\item{data}{Distance/drive time}

\item{kid_per_dot}{Number of kids per dot on map}
}
\value{
A tibble
}
\description{
Create synthetic locations for families within each geography
}
