% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_planar.R
\name{to_longlat}
\alias{to_longlat}
\title{Convert planar to long/lat coordinates}
\usage{
to_longlat(df)
}
\arguments{
\item{df}{sf tibble}
}
\value{
An sf tibble
}
\description{
Convert planar coordinates to long/lat coordinates.
}
