% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_planar.R
\name{to_planar}
\alias{to_planar}
\title{Convert long/lat to planar coordinates}
\usage{
to_planar(df)
}
\arguments{
\item{df}{sf tibble}
}
\value{
An sf tibble
}
\description{
Convert long/;at coordinates to planar coordinates.
}
