% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_locs.R
\name{add_locs}
\alias{add_locs}
\title{Add geographies}
\usage{
add_locs(data, gdrive_root = "~/Google Drive")
}
\arguments{
\item{data}{A tibble with address fields \code{street}, \code{city}, \code{state}.}

\item{gdrive_root}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
A tibble
}
\description{
Add administrative and other geographies into datasets.
}
