% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{get_drive_time}
\alias{get_drive_time}
\title{Get Euclidean and driving distance}
\usage{
get_drive_time(
  geography = "census-tract",
  year = 2010,
  destination = "licensing",
  date = "2021-04-01"
)
}
\arguments{
\item{geography}{Geographies to add to. Defaults to all geographies available.}

\item{year}{Vintage for the geography}

\item{destination}{List of providers to calculate distance to}

\item{date}{Date}
}
\value{
A tibble
}
\description{
Get (Euclidean) distance and driving duration using OSRM.
}
