% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot shapefiles}
\usage{
plot_map(data, method = "sf", fill_with)
}
\arguments{
\item{data}{An sf object that contains the geometries}

\item{method}{Plotting method. Can be "sf" or "tmap"}

\item{fill_with}{(Optional) The variable from the sf object to fill the polygons.}
}
\value{
A ggplot or tmap object.
}
\description{
Plot geography shapefiles using geom_sf
}
