% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licensing.R
\name{read_licensing}
\alias{read_licensing}
\title{Read DHS Licensing Data}
\usage{
read_licensing(table = "provider", version = "2021-04", filetype = "csv")
}
\arguments{
\item{table}{Licensing table to read. Either "provider", or "shift".}

\item{version}{Licensing data version. Either "2021-04" or "2021-02".}

\item{filetype}{Data file type. Either "csv" or "rds".}
}
\value{
A tibble of Licensing data.
}
\description{
Read DHS Licensing Data from the shared Google Drive folder.
}
