% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nware.R
\name{read_nware}
\alias{read_nware}
\title{Read Nware Data}
\usage{
read_nware(
  table = "characteristics",
  version = "2021-04",
  filetype = "rds",
  gdrive_root = "~/Google Drive"
)
}
\arguments{
\item{table}{Nware table to read. Either "characteristics", "provider_schedule_daily",  "provider_schedule_hourly", "shift", or "age_group_shift".}

\item{version}{Nware version. Either "2021-04" or "2020-10".}

\item{filetype}{Data file type. Either "csv" or "rds".}

\item{gdrive_root}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
A tibble of Nware data.
}
\description{
Read Nware Data from the shared Google Drive folder.
}
