% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parentaware.R
\name{read_parentaware}
\alias{read_parentaware}
\title{Read Parent Aware Data}
\usage{
read_parentaware(filetype = "rds", GDRIVE_ROOT = Sys.getenv("GDRIVE_ROOT"))
}
\arguments{
\item{filetype}{"rds" or "csv". Defaults to "rds".}

\item{GDRIVE_ROOT}{The local path to the folder that contains the MinnCCAccess folder. Defaults to "~/Google Drive".}
}
\value{
A tibble
}
\description{
Read Parent Aware Data from the shared Google Drive folder.
}
