% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pecc.R
\name{read_pecc}
\alias{read_pecc}
\title{Read PECC Grant Data}
\usage{
read_pecc(filetype = "rds", gdrive_root = "~/Google Drive")
}
\arguments{
\item{filetype}{"rds" or "csv"}

\item{gdrive_root}{The local path to the folder that contains the MinnCCAccess folder.}
}
\value{
A tibble of PECC grant data.
}
\description{
Read PECC Grant Data from the shared Google Drive folder.
}
