% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_public.R
\name{read_public}
\alias{read_public}
\title{Read Public School Enrollment Data}
\source{
Minnesota Department of Education
}
\usage{
read_public(
  table = "school",
  subgroups = FALSE,
  all_grades = FALSE,
  filetype = "rds",
  gdrive_root = "~/Google Drive"
)
}
\arguments{
\item{table}{"school" or "school-district". Defaults to "school".}

\item{subgroups}{Set TRUE to add demographic (gender, race) and other subgroups. Defaults to FALSE.}

\item{all_grades}{Set TRUE to get grade school enrollments as well as enrollment in early education. Defaults to FALSE.}

\item{filetype}{"rds" or "csv". Defaults to "rds".}

\item{gdrive_root}{The local path to the folder that contains the MinnCCAccess folder. Defaults to "~/Google Drive".}
}
\value{
A tibble
}
\description{
Read Public School Enrollment Data from the shared Google Drive folder.
}
