% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vpk.R
\name{read_vpk}
\alias{read_vpk}
\title{Load VPK data}
\usage{
read_vpk(source = "enrollment", geography = "census-place", geo_year = 2010)
}
\arguments{
\item{source}{"enrollment" or "allocation"}
}
\description{
Load VPK data using various sources
}
